requirements.txt
psutil==5.9.1
firebase_admin==5.2.0
pycryptodome==3.14.1
imutils==0.5.4
python-telegram-bot==13.12
opencv-python==4.5.1.48
